% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{analyse_rds}
\alias{analyse_rds}
\title{Get exports which don't have any examples}
\usage{
analyse_rds(path = ".")
}
\arguments{
\item{path}{Path to package}
}
\value{
Logical vector of whether exported functions have examples
}
\description{
Get exports which don't have any examples
}
\keyword{internal}
