% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_review.R
\name{analyse_vignette}
\alias{analyse_vignette}
\title{Analyse a single vignette}
\usage{
analyse_vignette(vig_path, checks = get_config()$vignettes)
}
\arguments{
\item{vig_path}{Path to directory where vignette is}
}
\description{
Analyse a single vignette
}
\keyword{internal}
