% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_functions.R
\name{check_assigned}
\alias{check_assigned}
\title{Check if code contains assignment}
\usage{
check_assigned(code, func_args)
}
\arguments{
\item{code}{Code to check}

\item{func_args}{Function argument to look for}
}
\description{
Check if code contains assignment
}
\examples{
# returns TRUE
check_assigned("x <- Array$create(c(2, 4, 6))")

# returns FALSE
check_assigned("Array$create(c(1,2), type = int8())")
}
\keyword{internal}
