% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_review.R
\name{check_image_alt_text}
\alias{check_image_alt_text}
\title{Check each image for an alt text description}
\usage{
check_image_alt_text(vig_path)
}
\arguments{
\item{vig_path}{Path to vignette}
}
\description{
Check each image for an alt text description
}
\keyword{internal}
