% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docreview.R
\name{check_results}
\alias{check_results}
\title{Check results and raise error if necessary}
\usage{
check_results(results, error_on_failure, error_on_warning)
}
\arguments{
\item{results}{List of docreview results}

\item{error_on_failure}{Raise error on review failures?}

\item{error_on_warning}{Raise error on review warnings?}
}
\description{
Inspired by testthat
}
\keyword{internal}
