% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_review.R
\name{divide_by_section}
\alias{divide_by_section}
\title{Divide by section}
\usage{
divide_by_section(md_text)
}
\arguments{
\item{md_text}{Markdown text}
}
\description{
Split a markdown document based on '<SECTION'> tags added a a previous stage
}
\keyword{internal}
