% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_examples}
\alias{find_examples}
\title{Get examples from a package}
\usage{
find_examples(path = ".")
}
\arguments{
\item{path}{Path to package}
}
\description{
Get examples from a package
}
\keyword{internal}
