% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_exports_without_examples}
\alias{find_exports_without_examples}
\title{Get exports which don't have any examples}
\usage{
find_exports_without_examples(path = ".")
}
\arguments{
\item{path}{Path to package}
}
\value{
Logical vector of whether exported functions have examples
}
\description{
Get exports which don't have any examples
}
\keyword{internal}
