% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_vignettes}
\alias{find_vignettes}
\title{Find vignettes}
\usage{
find_vignettes(path = ".")
}
\arguments{
\item{path}{Path to package}
}
\description{
Find all vignettes in a package
}
\keyword{internal}
