% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_review.R
\name{function_get_comments}
\alias{function_get_comments}
\title{Count failures and warnings for function review}
\usage{
function_get_comments(results, checks)
}
\arguments{
\item{results}{Results of function review}

\item{thresholds}{List of thresholds that result in fails or warnings}
}
\description{
Count failures and warnings for function review
}
\keyword{internal}
