% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{function_results_display}
\alias{function_results_display}
\title{Parse function documentation analysis results}
\usage{
function_results_display(results, thresholds)
}
\arguments{
\item{results}{Results of function documentation analysis}
}
\description{
Parse function documentation analysis results
}
\keyword{internal}
