% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{function_results_parse}
\alias{function_results_parse}
\title{Parse function documentation analysis results}
\usage{
function_results_parse(results, error_on_failure)
}
\arguments{
\item{results}{Results of function documentation analysis}

\item{error_on_failure}{Raise an error on any negative reviews}
}
\description{
Parse function documentation analysis results
}
\keyword{internal}
