% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docreview.R
\name{function_review}
\alias{function_review}
\title{Review functions}
\usage{
function_review(path, thresholds = default_thresholds()$functions)
}
\arguments{
\item{path}{Path to package}

\item{thresholds}{List of thresholds that result in fails or warnings}
}
\description{
Review functions
}
