% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rd.R
\name{get_aliases}
\alias{get_aliases}
\title{Get all alias tags from RD file}
\usage{
get_aliases(rd_path)
}
\arguments{
\item{rd_path}{Path to RD file}
}
\description{
Get all alias tags from RD file
}
\keyword{internal}
