% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exports_without_examples.R
\name{get_examples}
\alias{get_examples}
\title{Get examples from a package}
\usage{
get_examples(pkg_path)
}
\arguments{
\item{pkg_path}{Path to package}
}
\description{
Get examples from a package
}
