% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exports_without_examples.R
\name{get_exports}
\alias{get_exports}
\title{Get exports}
\usage{
get_exports(path)
}
\arguments{
\item{path}{Package root}
}
\value{
Exported functions, character vector
}
\description{
Get exports
}
