% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_functions.R
\name{get_exports_without_examples}
\alias{get_exports_without_examples}
\title{Get exports which don't have any examples}
\usage{
get_exports_without_examples(path)
}
\arguments{
\item{path}{Path to package}
}
\description{
Get exports which don't have any examples
}
