% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_vignettes.R
\name{get_fk_score}
\alias{get_fk_score}
\title{Get Flesch-Kincaid readability score}
\usage{
get_fk_score(code)
}
\arguments{
\item{code}{List of code chunks}
}
\description{
Get Flesch-Kincaid readability score
}
