% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rd.R
\name{get_function_args}
\alias{get_function_args}
\title{Get the arguments of all functions documented in this RD file}
\usage{
get_function_args(rd_path)
}
\arguments{
\item{rd_path}{Path to RD file}
}
\description{
Get the arguments of all functions documented in this RD file
}
\keyword{internal}
