% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_functions.R
\name{get_function_names}
\alias{get_function_names}
\title{Get the names of all functions documented in this RD file}
\usage{
get_function_names(rd_path)
}
\description{
Get the names of all functions documented in this RD file
}
