% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_review.R
\name{get_length}
\alias{get_length}
\title{Get number of words in a markdown chunk}
\usage{
get_length(md)
}
\arguments{
\item{md}{Markdown chunk}
}
\description{
Get number of words in a markdown chunk
}
\keyword{internal}
