% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_functions.R
\name{get_rd_filename_for_function}
\alias{get_rd_filename_for_function}
\title{Get the RD file containing documentation for a function}
\usage{
get_rd_filename_for_function(pkg_path, function_name)
}
\description{
Get the RD file containing documentation for a function
}
