% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_functions.R
\name{parse_example}
\alias{parse_example}
\title{Parse a function call and divide up by parameters}
\usage{
parse_example(expr_text, func_name)
}
\arguments{
\item{expr_text}{Function call expression text}

\item{func_name}{Function name}
}
\description{
Take a function call and split it up so we have each parameter expression separate
Apparently bracket matching can't be done using regex so this manual approach is needed
}
\examples{
# First param is func call
parse_example("match_arrow(Scalar$create(\"Mazda RX4 Wag\"), cars_tbl$name)", "match_arrow")

# Can handle inline comments
parse_example("match_arrow(Scalar$create(4), cars_tbl$cyl) # 0-indexed", "match_arrow")

# First param is func call containing vector
parse_example("match_arrow(Array$create(c(4, 6, 8)), cars_tbl$cyl)", "match_arrow")

# Second param is func call containing vector
parse_example("match_arrow(cars_tbl$cyl, Array$create(c(4, 6, 8)))", "match_arrow")
}
