% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_vignettes.R
\name{parse_vignette}
\alias{parse_vignette}
\title{Parse a vignette}
\usage{
parse_vignette(vig_path)
}
\arguments{
\item{vig_path}{Path to vignette}
}
\value{
List of length 1 character vectors containing contents of each markdown section
}
\description{
Parse a vignette
}
\keyword{internal}
