% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{parse_vignette_results}
\alias{parse_vignette_results}
\title{Parse Vignette Results}
\usage{
parse_vignette_results(vignette_results)
}
\arguments{
\item{vignette_results}{Output of calling vignette analysis function}
}
\description{
Parse Vignette Results
}
