% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_review.R
\name{remove_backticks}
\alias{remove_backticks}
\title{Remove any code in backticks from a chunk}
\usage{
remove_backticks(chunk)
}
\arguments{
\item{chunk}{Code chunk}
}
\description{
Remove any code in backticks from a chunk
}
\keyword{internal}
