% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_review.R
\name{remove_bullets}
\alias{remove_bullets}
\title{Remove bullet points from a chunk}
\usage{
remove_bullets(chunk)
}
\arguments{
\item{chunk}{Code chunk}
}
\description{
Remove any bullet points in markdown format
}
\keyword{internal}
