% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_review.R
\name{remove_code}
\alias{remove_code}
\title{Remove any code in backticks from a chunk}
\usage{
remove_code(chunk)
}
\arguments{
\item{chunk}{Code chunk}
}
\description{
Remove any code in backticks from a chunk
}
\keyword{internal}
