% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_review.R
\name{remove_pipe_tables}
\alias{remove_pipe_tables}
\title{Remove pipe table from a chunk}
\usage{
remove_pipe_tables(chunk)
}
\arguments{
\item{chunk}{Code chunk}
}
\description{
Remove any pipe tables in markdown code
}
\keyword{internal}
