% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_review.R
\name{set_thresholds}
\alias{set_thresholds}
\title{Set thresholds at which review checks should fail}
\usage{
set_thresholds(
  exports_without_examples = "fail",
  fk_fail = 30,
  fk_warn = 50,
  length_fail = 3000,
  length_warn = 2000
)
}
\arguments{
\item{exports_without_examples}{Action to take when exports without examples
are identified. Possible values are "fail", "warn", or "none".}

\item{fk_fail}{Vignette Flesch-Kincaid scores lower than this will result in
a review check failure}

\item{fk_warn}{Vignette Flesch-Kincaid scores lower than this but higher than
\code{fk_fail} will results in a review check warning}

\item{length_fail}{Vignette word count longer than this will result in a
review check failure}

\item{length_warn}{Vignette word count longer than this but shorter than
\code{length_fail} will result in a review check warning}
}
\description{
lifecycle::badge('experimental')
This function is experimental and as the key package metrics evolve,
parameters may be subject to change.
}
\examples{
set_thresholds(exports_without_examples = "warn")
}
