% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docreview.R
\name{vignette_review}
\alias{vignette_review}
\title{Review vignettes}
\usage{
vignette_review(path, thresholds = default_thresholds()$vignettes)
}
\arguments{
\item{path}{Path to package}
}
\description{
Review vignettes
}
