% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_functions.R
\name{check_assigned}
\alias{check_assigned}
\title{returns TRUE

check_assigned("x <- Array$create(c(2, 4, 6))")}
\usage{
check_assigned(code, func_args)
}
\description{
returns TRUE

check_assigned("x <- Array$create(c(2, 4, 6))")
}
\section{returns FALSE}{
check_assigned("Array$create(c(1,2), type = int8())")
}

