% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_review.R
\name{check_results}
\alias{check_results}
\title{Check results and raise error if necessary}
\usage{
check_results(results, config)
}
\arguments{
\item{results}{List of docreview results}

\item{config}{List containing logical values \code{error_on_warning} and \code{error_on_failure}}
}
\description{
Inspired by testthat
}
\keyword{internal}
