% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_review.R
\name{clean_chunks}
\alias{clean_chunks}
\title{Clean markdown chunks}
\usage{
clean_chunks(chunk)
}
\arguments{
\item{chunk}{Code chunk}
}
\description{
Remove text in links and backticks from chunks
}
\keyword{internal}
