% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docreview.R
\docType{package}
\name{docreview-package}
\alias{docreview}
\alias{docreview-package}
\title{docreview: Opiniated Documentation Checking}
\description{
docreview: reviewing package documentation
}
\author{
\strong{Maintainer}: Nic Crane \email{thisisnic@gmail.com}

}
\keyword{internal}
