% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_vignettes.R
\name{extract_md_section}
\alias{extract_md_section}
\title{Extract markdown section}
\usage{
extract_md_section(md)
}
\arguments{
\item{md}{Markdown}
}
\description{
Extract the content from the markdown section and collapse it into one string
}
