% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_review.R
\name{function_review}
\alias{function_review}
\title{Review functions}
\usage{
function_review(path, checks = get_config()$functions)
}
\arguments{
\item{path}{Path to package}

\item{checks}{Checks to run}
}
\description{
Review functions
}
