% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_review.R
\name{function_review}
\alias{function_review}
\title{Review functions}
\usage{
function_review(path, thresholds = default_thresholds()$functions)
}
\arguments{
\item{path}{Path to package}

\item{thresholds}{List of thresholds that result in fails or warnings}
}
\description{
Review functions
}
\examples{
pkg_path <- system.file("testpkg", package = "docreview")
function_review(pkg_path)
}
