% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_functions.R
\name{get_function_args}
\alias{get_function_args}
\title{Get the arguments of all functions documented in this RD file}
\usage{
get_function_args(rd_path)
}
\description{
Get the arguments of all functions documented in this RD file
}
