% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docreview.R
\name{package_review}
\alias{package_review}
\title{Review package documentation}
\usage{
package_review(
  path = ".",
  error_on_failure = FALSE,
  review = c("functions", "vignettes")
)
}
\arguments{
\item{path}{Path to package}
}
\description{
Review package documentation
}
