% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docreview.R
\name{package_review}
\alias{package_review}
\title{Review package documentation}
\usage{
package_review(
  path = ".",
  error_on_failure = FALSE,
  error_on_warning = FALSE,
  doc_types = c("functions", "vignettes"),
  thresholds = default_thresholds()
)
}
\arguments{
\item{path}{Path to package}

\item{error_on_failure}{Raise an error on any negative reviews}

\item{doc_types}{Types of documentation to review}

\item{thresholds}{List of thresholds that result in fails or warnings}
}
\description{
Review package documentation
}
