% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docreview.R
\name{package_review}
\alias{package_review}
\title{Review package documentation}
\usage{
package_review(
  path = ".",
  error_on_failure = FALSE,
  doc_types = c("functions", "vignettes")
)
}
\arguments{
\item{path}{Path to package}

\item{error_on_failure}{Raise an error on any negative reviews}

\item{doc_types}{Types of documentation to review}
}
\description{
Review package documentation
}
