% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_vignettes.R
\name{remove_links}
\alias{remove_links}
\title{Remove links from a chunk}
\usage{
remove_links(chunk)
}
\arguments{
\item{chunk}{Code chunk}
}
\description{
Remove any links in markdown format
}
\keyword{internal}
