% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_review.R
\name{set_thresholds}
\alias{set_thresholds}
\title{Set thresholds at which review checks should fail}
\usage{
set_thresholds(
  values = list(exports_without_examples = "fail", fk_fail = 30, fk_warn = 50,
    length_fail = 3000, length_warn = 2000)
)
}
\arguments{
\item{values}{List of values:
\itemize{
\item{exports_without_examples} - {Action to take when exports without
examples are identified. Possible values are "fail", "warn", or "none".}
}}
}
\description{
Set thresholds at which review checks should fail
}
\examples{
set_thresholds()
}
