% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{vignette_results_display}
\alias{vignette_results_display}
\title{Parse Vignette Results}
\usage{
vignette_results_display(results, config)
}
\arguments{
\item{config}{list containing config settings}

\item{vignette_results}{Output of calling vignette analysis function}
}
\description{
Parse Vignette Results
}
\keyword{internal}
