% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{vignette_results_parse}
\alias{vignette_results_parse}
\title{Parse Vignette Results}
\usage{
vignette_results_parse(vignette_results, error_on_failure)
}
\arguments{
\item{vignette_results}{Output of calling vignette analysis function}

\item{error_on_failure}{Raise an error on any negative reviews}
}
\description{
Parse Vignette Results
}
