% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_review.R
\name{vignette_review}
\alias{vignette_review}
\title{Review vignettes}
\usage{
vignette_review(path, checks = get_config()$vignettes)
}
\arguments{
\item{path}{Path to package}

\item{checks}{Checks to run}
}
\description{
Review vignettes
}
