% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docreview.R
\name{vignette_review}
\alias{vignette_review}
\title{Review vignettes}
\usage{
vignette_review(path, thresholds = default_thresholds()$vignettes)
}
\arguments{
\item{path}{Path to package}

\item{thresholds}{List of thresholds that result in fails or warnings}
}
\description{
Review vignettes
}
\examples{
pkg_path <- system.file("testpkg", package = "docreview")
vignette_review(pkg_path)
}
